% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedSplineModel.R
\docType{class}
\name{FittedSplineModel-class}
\alias{FittedSplineModel-class}
\alias{FittedSplineModel}
\title{A Spline Model}
\value{
an instance of \code{\link{FittedSplineModel}}
}
\description{
This class holds a spline model.
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the spline approach}
}}

\seealso{
FittedSplineModel.new
}
