% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedSplineModel.R
\name{FittedSplineModel.as.character}
\alias{FittedSplineModel.as.character}
\title{Convert a \code{FittedSplineModel} to a String}
\usage{
FittedSplineModel.as.character(x)
}
\arguments{
\item{x}{the \code{FittedSplineModel}}
}
\value{
the string
}
\description{
well, convert a \code{FittedSplineModel} to a String
}
