% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedSplineModel.R
\name{FittedSplineModel.new}
\alias{FittedSplineModel.new}
\title{Create a New Instance of \code{\link{FittedSplineModel}}.}
\usage{
FittedSplineModel.new(f, quality, size, name)
}
\arguments{
\item{f}{a function accepting one parameter and returning a value}

\item{quality}{the quality of the model on the original data, computed by a
quality metric, smaller values are better}

\item{size}{the size of the model, i.e., the number of parameters}

\item{name}{the name of the spline model}
}
\value{
the new instance
}
\description{
Create a New Instance of \code{\link{FittedSplineModel}}.
}
