% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{regressoR.spline.default}
\alias{regressoR.spline.default}
\title{Get the Default Spline Fitters}
\usage{
regressoR.spline.default()
}
\value{
the list of default splines
}
\description{
Get the default fitters for splines.
}
