% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{regressoR.spline.protected}
\alias{regressoR.spline.protected}
\title{Get the Default Protected Spline Fitters}
\usage{
regressoR.spline.protected()
}
\value{
the list of default protected splines
}
\description{
Get the default fitters for protected splines, i.e., splines
  which will remain constant before and after the first and last data point,
  respectively.
}
