% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothSpline.R
\name{regressoR.spline.smooth}
\alias{regressoR.spline.smooth}
\title{Fit a Smoothed Spline to a given Dataset}
\usage{
regressoR.spline.smooth(metric, transformation.x = NULL,
  transformation.y = NULL, metric.transformed = NULL, protected = TRUE,
  ...)
}
\arguments{
\item{metric}{an instance of \code{RegressionQualityMetric}}

\item{transformation.x}{the transformation along the \code{x}-axis, or
\code{NULL} if none was applied to the data}

\item{transformation.y}{the transformation along the \code{y}-axis, or
\code{NULL} if none was applied to the data}

\item{metric.transformed}{the transformed metric for the first fitting step}

\item{protected}{should the function be limited to the range of values
actually ocurring in the data?}

\item{...}{parameters to be passed to \code{\link[stats]{smooth.spline}}}
}
\value{
On success, an instance of \code{\link{FittedSplineModel}},
  \code{NULL} on failure.
}
\description{
A smooth spline is fitted to the given dataset.
}
